combine file plugin integration tests
