+++ app/secrets/secret_test.go
@@ -5,6 +5,8 @@ import (
 	"crypto/aes"
 	"crypto/cipher"
 	"encoding/base64"
+	"os"
+	"path/filepath"
 	"sync"
 	"testing"
 
@@ -112,5 +114,42 @@ func TestLoadSecretAWSKMS(t *testing.T) {
 	}
 	if val != "secret" {
 		t.Fatalf("expected 'secret', got %s", val)
+	}
+}
+
+func TestLoadSecretFile(t *testing.T) {
+	dir := t.TempDir()
+	path := filepath.Join(dir, "secret.txt")
+	if err := os.WriteFile(path, []byte("top-secret"), 0600); err != nil {
+		t.Fatalf("write: %v", err)
+	}
+	val, err := secrets.LoadSecret(context.Background(), "file:"+path)
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+	if val != "top-secret" {
+		t.Fatalf("expected 'top-secret', got %s", val)
+	}
+}
+
+func TestLoadSecretFileTrailingNewline(t *testing.T) {
+	dir := t.TempDir()
+	path := filepath.Join(dir, "secret.txt")
+	if err := os.WriteFile(path, []byte("top-secret\n"), 0600); err != nil {
+		t.Fatalf("write: %v", err)
+	}
+	val, err := secrets.LoadSecret(context.Background(), "file:"+path)
+	if err != nil {
+		t.Fatalf("unexpected error: %v", err)
+	}
+	if val != "top-secret" {
+		t.Fatalf("expected 'top-secret', got %s", val)
+	}
+}
+
+func TestLoadSecretFileMissing(t *testing.T) {
+	path := filepath.Join(t.TempDir(), "missing.txt")
+	if _, err := secrets.LoadSecret(context.Background(), "file:"+path); err == nil {
+		t.Fatal("expected error for missing file")
 	}
 }
